Worked on finishing hazards
added enum LASTMOVE to track last move of car to prevent infinite loops while avoiding hazards

Added code to findMove to check if the car is going to move onto  a visible HAZARD after finding a target. If it does, it does a coin flip to determine which direction on the other axis to move to avoid it (if moving left/right onto a hazard, move up/down. If moving up/down, move left/right)